requirements.txt
fastapi
uvicorn
python-dotenv
dhanhq
jinja2
openai
kiteconnect